Grind main/menu_equip EquipCtrlCur: (s16)(u16) pad cast match (misalign 10->8) (#17533)
rawPress/rawHold were u16 with (s16) cast emitting a bare extsh; the target
materializes the u16 in a 32-bit reg and emits clrlwi+extsh. Retype to
unsigned int and cast (s16)(u16) to reproduce the mask-then-sign-extend,
aligning 2 instructions. Unit fuzzy 95.86 -> 95.94, no regressions.

+++ src/menu_equip.cpp
@@ -298,7 +298,7 @@ int CMenuPcs::EquipCtrlCur()
 	if ((padLock != 0) || (Pad.m_debugPadPort != -1)) {
 		blocked = true;
 	}
-	u16 rawPress;
+	unsigned int rawPress;
 	if (blocked) {
 		rawPress = 0;
 	} else {
@@ -306,13 +306,13 @@ int CMenuPcs::EquipCtrlCur()
 		padIndex &= ~-((__cntlzw((unsigned int)Pad.m_debugPadPort) & 0x20) >> 5);
 		rawPress = Pad.GetPadInputs()[padIndex].buttonDown[0];
 	}
-	press = (s16)rawPress;
+	press = (s16)(u16)rawPress;
 
 	blocked = false;
 	if ((padLock != 0) || (Pad.m_debugPadPort != -1)) {
 		blocked = true;
 	}
-	u16 rawHold;
+	unsigned int rawHold;
 	if (blocked) {
 		rawHold = 0;
 	} else {
@@ -320,7 +320,7 @@ int CMenuPcs::EquipCtrlCur()
 		padIndex &= ~-((__cntlzw((unsigned int)Pad.m_debugPadPort) & 0x20) >> 5);
 		rawHold = Pad.GetPadInputs()[padIndex].repeatButton;
 	}
-	hold = (s16)rawHold;
+	hold = (s16)(u16)rawHold;
 
 	if (hold == 0) {
 		return 0;
